+++ av/video/stream.py
@@ -40,7 +40,7 @@ class VideoStream(Stream):
 
         .. seealso:: This is mostly a passthrough to :meth:`.CodecContext.encode`.
         """
-
+        self._assert_has_codec_context(lib.AVERROR_ENCODER_NOT_FOUND)
         packets = self.codec_context.encode(frame)
         packet: Packet
         for packet in packets:
@@ -57,6 +57,7 @@ class VideoStream(Stream):
 
         .. seealso:: This is a passthrough to :meth:`.CodecContext.decode`.
         """
+        self._assert_has_codec_context()
         return self.codec_context.decode(packet)
 
     @cython.cfunc